scripts: add auto-options.lua
This replaces the old ugly 2-configs-with-one-symlinked approach with a
new script-sets-options-dynamically approach, which is nearly as ugly,
since it basically mirrors the config file (well, parts of it).
At some point mpv might support setting profiles with lua scripts, then
this script will be significantly shorter. The appropriate profiles are
already defined in mpv.conf.